Refactor: separate module css
