+++ src/components/CreateRollingPaperPage/PickBackground.module.css
@@ -1,0 +1,23 @@
+h1 {
+  margin: 14px 0;
+  font-size: 24px;
+}
+
+p {
+  margin: 12px 0;
+  color: var(--gray-500);
+}
+
+.toggleButton {
+  width: 122px;
+  height: 40px;
+  border: none;
+  border-radius: 6px;
+  font-weight: bold;
+}
+
+.toggleButtonOn {
+  border: 2px solid var(--purple-600);
+  background-color: var(--white);
+  color: var(--purple-700);
+}
