+++ src/pages/CreateRollingPaperPage.module.css
@@ -1,105 +1,7 @@
-main {
+.container {
   display: flex;
   width: 720px;
   flex-direction: column;
   margin: 122px auto;
   gap: 50px;
-}
-
-h1 {
-  margin: 14px 0;
-  font-size: 24px;
-}
-
-p {
-  margin: 12px 0;
-  color: var(--gray-500);
-}
-
-.pickerTextContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.pickerContainer {
-  display: flex;
-  width: 720px;
-  height: 168px;
-  gap: 16px;
-}
-
-.pickerBox {
-  position: relative;
-  display: flex;
-  width: 168px;
-  height: 168px;
-  align-items: center;
-  justify-content: center;
-  border-radius: 8%;
-  cursor: pointer;
-}
-
-.beige {
-  background-color: var(--orange-200);
-}
-
-.purple {
-  background-color: var(--purple-200);
-}
-
-.blue {
-  background-color: var(--blue-200);
-}
-
-.green {
-  background-color: var(--green-200);
-}
-
-.toggleButton {
-  width: 122px;
-  height: 40px;
-  border: none;
-  border-radius: 6px;
-  font-weight: bold;
-}
-
-.toggleButtonOn {
-  border: 2px solid var(--purple-600);
-  background-color: var(--white);
-  color: var(--purple-700);
-}
-
-.createButton {
-  height: 56px;
-  border: none;
-  border-radius: 12px;
-  background-color: var(--purple-600);
-  color: var(--white);
-  font-size: 18px;
-  font-weight: bold;
-  opacity: 0.3;
-}
-
-.activateButton {
-  opacity: 1;
-}
-
-img {
-  width: 168px;
-  height: 168px;
-  border-radius: 8%;
-  object-fit: cover;
-}
-
-.check {
-  position: absolute;
-  left: 50%;
-  width: 44px;
-  height: 44px;
-  transform: translateX(-50%);
-}
-
-.imagesOpacity {
-  opacity: 0.3;
 }
